removing unnessary file and fixing errors
