+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: amazon_s3_cognito
 description: This plugin allows users to list files, upload and delete image for amazon s3 cognito
-version: 0.6.0
+version: 0.6.2
 homepage: https://github.com/prachiFam/amazon_s3_cognito
 
 environment:
